+++ backend/requirements.txt
@@ -2,6 +2,7 @@ aiohttp==3.9.5
 aiosignal==1.3.1
 annotated-types==0.7.0
 anyio==4.4.0
+APScheduler==3.10.4
 attrs==23.2.0
 beautifulsoup4==4.12.3
 blinker==1.7.0
@@ -122,7 +123,9 @@ typer==0.9.4
 typing-inspect==0.9.0
 typing_extensions==4.12.0
 tzdata==2024.1
+tzlocal==5.2
 urllib3==2.2.0
+uuid==1.30
 wasabi==1.1.3
 weasel==0.3.4
 Werkzeug==3.0.1
